backend/app/requirements.txt

breadability==0.1.20
brotlipy==0.7.0
constantly==15.1.0
cssselect==1.1.0
uvicorn==0.20.0
cytoolz==0.11.0
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.95.0
fonttools==4.25.0
h11==0.14.0
inflection==0.5.1
llvmlite==0.38.0
beautifulsoup4==4.11.1
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
requests==2.28.1
opencv-python==4.7.0.72
openpyxl==3.0.10
outcome==1.2.0
pandas==1.4.4
patsy==0.5.2
pep8==1.7.1
Pillow==9.2.0
ply==3.11
psycopg2==2.9.5
nltk==3.7
pyasn1-modules==0.2.8
pycosat==0.6.3
pycountry==22.3.5
pycurl==7.45.1
pydantic==1.10.6
PyDispatcher==2.0.5
pyls-spyder==0.4.0
PyQt5-sip==12.11.0
pytest==7.1.2
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
queuelib==1.5.0
seaborn==0.12.0
selenium==4.8.2
SQLAlchemy==1.4.39 
starlette==0.26.1
sumy==0.11.0
trio==0.22.0
trio-websocket==0.10.2
webdriver-manager==3.8.5
webencodings==0.5.1
wordcloud==1.8.2.2
wsproto==1.2.0
zict==2.1.0
webdriver-manager==3.8.5
#libasound2==1.0.17
#libatk-bridge==2.5.3
#libatk1==2.2.0
#0 2.235                         Depends: libatspi2.0-0 (>= 2.9.90) but it is not installable
#0 2.235                         Depends: libcups2 (>= 1.6.0) but it is not installable
#0 2.235                         Depends: libdbus-1-3 (>= 1.9.14) but it is not installable
#0 2.235                         Depends: libdrm2 (>= 2.4.75) but it is not installable
#0 2.235                         Depends: libgbm1 (>= 17.1.0~rc2) but it is not installable
#0 2.235                         Depends: libgtk-3-0 (>= 3.9.10) but it is not installable or
#0 2.235                                  libgtk-4-1 but it is not installable
#0 2.235                         Depends: libnspr4 (>= 2:4.9-2~) but it is not installable
#0 2.235                         Depends: libnss3 (>= 2:3.26) but it is not installable
#libu2f-udev but it is not installable
#libvulkan1 but it is not installable
#libxcomposite1 (>= 1:0.4.4-1) but it is not installable
#libxdamage1 (>= 1:1.1) but it is not installable
#libxfixes3 but it is not installable
#libxkbcommon0==0.5.0
#libxrandr2==2.0.0 
#xdg-utils==1.0.2
